created routes and added users/expenses there, added extensions for bcrypt
