+++ spending_tracker/extensions.py
@@ -1,0 +1,2 @@
+from flask_bcrypt import Bcrypt
+bcrypt = Bcrypt()
